:art: Move the Snek into it's own Snekky Folder
