+++ frontend/src/components/Snek/SnekN.js
