+++ frontend/src/components/Snek/SnekStart.js
