+++ frontend/src/components/Snek/index.js
@@ -1,6 +1,5 @@
 import React from "react";
 import PropTypes from "prop-types";
-import key from "weak-key";
 import Snektion from "./Snektion";
 
 const Snek = ({ size, length }) => {
@@ -8,6 +7,7 @@ const Snek = ({ size, length }) => {
   for (var i = 0; i < length; i++) {
     snektion_array.push(<Snektion position={i} key={i} />)
   }
+
   return (
     <div className={"snek snek--" + size }>
     { snektion_array }
