+++ frontend/src/components/Table.js
@@ -8,18 +8,18 @@ const Table = ({ data }) =>
   ) : (
     <div className="column">
       <h2 className="subtitle">
-        Showing <strong>{data.length} items</strong>  
+        Showing <strong>{data.length} items</strong>
       </h2>
       <table className="table table--striped">
         <thead>
           <tr>
-            {Object.entries(data[0]).map(el => <th key={key(el)}>{el[0]}</th>)}
+            {Object.entries(data[0]).map((el, i) => <th key={i}>{el[0]}</th>)}
           </tr>
         </thead>
         <tbody>
-          {data.map(el => (
-            <tr key={el.id}>
-              {Object.entries(el).map(el => <td key={key(el)}>{el[1]}</td>)}
+          {data.map((el, i) => (
+            <tr key={i}>
+              {Object.entries(el).map((el, i) => <td key={i}>{el[1]}</td>)}
             </tr>
           ))}
         </tbody>
